feat: Add Action Buttons in attendance records column definition
